Dictionaries refactor
